+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release 5.5.0:
  - Update implementation of [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html) to draft 15:
    - Remove functionality to request issuance of certain claims only, as this has been dropped from OpenID4VCI entirely
    - Remove format-specific parameters in credential request, replacing with `credential_configuration_id`
+   - In credential response, replace single `credential` with array `credentials`, containing the `credential` itself, but issue both variants for now
    - Deprecate `WalletService.RequestOptions.requestedAttributes`
    - Deprecate methods in `OpenId4VciClient` containing parameter for `requestedAttributes`
    - In `OpenId4VciClient.startProvisioningWithAuthRequest()` remove parameter `requestedAttributes`